+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
   imports.
 - Raise at compile time when a build macro is called in a module that does not
   `use Doggo.Components`.
+- Raise an `ArgumentError` for an invalid `expanded` value in `accordion`.
 
 ## [0.16.0] - 2026-09-22